fix bug

+++ ConfigLoader.py
@@ -6,7 +6,7 @@ class ConfigLoader:
         with open(filename, 'r') as f:
             config = json.load(f)
         self.mode = config['mode']
-        self.dt_sec = int(1 / config['FPS'])
+        self.dt_sec = 1 / float(config['FPS'])
         self.port_stage = f'{config["PORT-stage"]}'
         self.port_laser = f'{config["PORT-laser"]}'
         self.baudrate_laser = config["BAUDRATE-laser"]
